pyproject.toml
[tool.poetry]
name = "sharp-meme-bot"
version = "0.1.0"
description = "Telegram-бот для мемних передбачень"
authors = ["Твоє ім'я <email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-telegram-bot = "^20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
